+++ src/backend/services/igdb/index.js
@@ -1,47 +1,14 @@
 import { path } from 'ramda';
 import { getGames, getGameById, getGenreById, getGameImagesByGameId, getGameLinksByGameId } from './igdb';
+import { constructGetGameOptions, constructSearchOptions, constructGetGameLinkOptions, constructGetGameImageOptions } from './igdb.config';
 
-const serverMethodOptions = {
-  callback: false,
-  cache: {
-    cache: 'diskCache',
-    expiresIn: 30 * 1000,
-    generateTimeout: 10000,
-  },
-};
 const register = (server, options, next) => {
-  const { Game, Game_Links, Game_Images } = path(['plugins', 'datastore', 'DatabaseModels'])(server);
-  server.method('search', getGames, { ...serverMethodOptions,
-    generateKey: params => (`${params.term || ''}${params.zone || ''}`) });
-  const geGameByIdOptions = {
-    ...serverMethodOptions,
-    bind: {
-      models: {
-        Game, Game_Links, Game_Images,
-      },
-    },
-  };
-  const getImageOptions = {
-    ...serverMethodOptions,
-    bind: {
-      models: {
-        Game_Images,
-      },
-    },
-  };
-
-  const getLinkOptions = {
-    ...serverMethodOptions,
-    bind: {
-      models: {
-        Game_Links,
-      },
-    },
-  };
-  server.method('getGameById', getGameById, geGameByIdOptions);
-  server.method('getGenreGameById', getGenreById, serverMethodOptions);
-  server.method('getGameImagesByGameId', getGameImagesByGameId, getImageOptions);
-  server.method('getGameLinksByGameId', getGameLinksByGameId, getLinkOptions);
+  const models = path(['plugins', 'datastore', 'DatabaseModels'])(server);
+  server.method('search', getGames, constructSearchOptions());
+  server.method('getGameById', getGameById, constructGetGameOptions(models));
+  server.method('getGenreGameById', getGenreById, constructGetGameOptions(models));
+  server.method('getGameImagesByGameId', getGameImagesByGameId, constructGetGameImageOptions(models));
+  server.method('getGameLinksByGameId', getGameLinksByGameId, constructGetGameLinkOptions(models));
   return next();
 };
 register.attributes = {
